Wave 44: 3725 -> 3777 (+52), bytes 53.86% -> 54.74%
Nine agents, 5-6 functions each on address-adjacent blocks, slots refilled
in address order. 52 new functions; both builds reproduce the ROM.

THE WAVE'S MEASUREMENT: assigned batches landed 44/54 (81%), extension work
8/19 (42%). Same agents, same blocks -- the difference is that an assigned
batch carries a shape the orchestrator read and a named trap, while extension
work carries a name and a path on an agent that has already spent its context.
Refilling a slot costs one orchestrator turn and beats writing longer
extension lists. Do not read the gap as agent quality.

PAIR VERIFICATION PAID, INCLUDING THE NEGATIVE. Diffing instruction streams
before briefing found three real pairs, and both members landed every time:
sub_0806377C/sub_080637D8 (22 insns, one differing `bl`), sub_08067504/
sub_0806752C (guard sense + callee), sub_0807FC70/sub_08080A70 (four
constants). The rejection mattered as much: sub_08061E98/sub_08061F34 share
an exemplar, sit adjacent and both carry loops, but are 72 vs 112 instructions
-- asserting that pair would have cost both functions.

REFUTED, MINE: the brief told W44-H that sub_08061A40 is agbcc's `movstrsi`
expansion of `*dst = *src` and to find the struct size. It is not. agbcc has
exactly two block-move paths and NEITHER emits an unrolled byte copy --
`movstrqi` is gated on align==4 && size<=48 and moves 12 bytes via ldm/stm
(48B -> four ldm/stm pairs, 52B -> `bl memcpy`, 0x130 -> `bl memcpy`), and
`move_by_pieces` is switched off by MOVE_RATIO (memcpy is a libcall at N=2,
3, 4, 8, 16 alike). No struct type can present alignment 1 to either path.
The per-file compiler axis was tested in four configurations and is negative.
Parked with all four axes closed. New "Block moves" chapter.

SELF-RELOCATING CODE, a class the project had not recorded. Three functions
at 0x0808AE54/AF00/AF74 copy a sibling's MACHINE CODE onto the stack and call
it there: THUMB-bit XOR to get a readable byte address, size as
`(int)nextFn - (int)copiedFn`, and `lsls #0xf; lsrs #0x10` as a (u16)(size/2)
halfword counter. Idiom fully derived and written up. All three parked on ONE
shared cause -- agbcc hoists `mov r3, sp; adds r3, #1` into the loop preheader
where LICM writes, keeping it live and shifting the whole function one
register left; the loop-test duplication is downstream of it. Control: with no
loop at all a probe puts `sp + 1` exactly where the ROM has it. Seven axes
closed; cross-referenced so the next wave takes them as one job.

sub_08063614 is ASM-RESIDENT, on two independent tells plus a closing scan.
It reads the PROGRAM COUNTER as a value (`mov r2, pc; lsrs #0x18`) to pick a
delay constant per memory region, and its loop is `subs; bgt` with no `cmp` --
W44-A measured that agbcc ALWAYS re-emits the compare and never reuses a
sub's own flags, so no C spelling exists. Orchestrator scan: exactly one
function ROM-wide reads pc as data, and no matched function does. Class closed
at one member.

RETYPES, all byte-neutral and verified by unit, not by caller: sub_0803CB24
`int` -> `u8` (both call sites test with a bare `lsls #0x18`, which an `int`
return cannot produce), sub_0808606C's parameter to its own struct, and
sub_08080F54's parameter widened. 29 promoted functions across 4 units --
including the 25-function c_0803C354.c -- re-verified clean. W44-G also found
a promoted caller passing one argument to the two-argument sub_0808AF00;
c_0801B598.c's whole unit re-verified 6/6.

17 new chapters in docs/agbcc-codegen.md. Among them W44-E's correction that
the wave-37 `ldrsh` rule does NOT carry over to QImode -- `lsls #0x18;
asrs #0x18` on a byte load genuinely IS an `(s8)` cast -- and W44-F's finding
that one of eleven callers of sub_0803CAB8 does not narrow its result.

Parked 4 with evidence. Deliberately did NOT park sub_08061DCC, sub_080723DC
(88.3%) or sub_080726E8: they are unfinished attempts with no ruled-out axes,
and parking those suppresses them from every future pre-flight while teaching
nothing. They stay in the queue; the 88.3% will resurface via best.json.

+++ src/decomp/c_0803CB24.c
@@ -14,7 +14,12 @@
  * sub_0803CB24 @ 0x0803CB24
  */
 
-int sub_0803CB24(u32 id)
+/* Return re-typed from `int` to `u8` in wave 44 (W44-C) -- byte-identical here
+ * (the `ldrb`-derived operand already has nonzero_bits <= 0xff, so the
+ * narrowing is a no-op), but sub_08084938 tests the result with a bare
+ * `lsls r0, r0, #0x18` at both call sites, which an `int` return cannot
+ * produce. See include/unknown-functions.h next to sub_0803CAD4. */
+u8 sub_0803CB24(u32 id)
 {
     struct Unk02028030 *s = &gUnknown_02028030;
     u32 idx = id >> 3;
